+++ trilby-cli/app/Main.hs
@@ -1,14 +1,11 @@
 module Main where
 
-import Control.Monad.Logger
-import Data.Maybe (fromMaybe)
-import Log (withLog)
-import Options.Applicative (execParser, (<|>))
-import System.Environment (lookupEnv)
+import Options.Applicative (execParser)
 import Trilby.App
 import Trilby.Command
 import Trilby.Install (install)
 import Trilby.Install.Options (validateParsedInstallOpts)
+import Trilby.Log (withLog)
 import Trilby.Options
 import Trilby.Update (update)
 import Prelude
@@ -16,10 +13,8 @@ import Prelude
 main :: IO ()
 main = do
     opts <- execParser parseOptionsInfo
-    state <- do
-        defaultVerbosity <- (readLogLevel =<<) <$> lookupEnv "TRILBY_VERBOSITY"
-        let verbosity = fromMaybe LevelWarn $ opts.verbosity <|> defaultVerbosity
-        pure AppState{..}
+    verbosity <- getVerbosity opts
+    let state = AppState{..}
     withLog state.verbosity do
         runApp state do
             case opts.command of