+++ trilby-cli/CHANGELOG.md
@@ -4,6 +4,7 @@
 
 * Introduce [CalVer](https://calver.org/) versioning (YY.MM.MICRO)
 * Add `--version` flag
+* Replace `--verbosity` with `--quiet`, `--verbose`, and `--debug`
 ## 0.1.0.0 -- YYYY-mm-dd
 
 * First version. Released on an unsuspecting world.